Removed migrations

+++ products/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Product(models.Model):
     name = models.CharField(max_length=120)
     image = models.ImageField(upload_to='products', default='no_picture.png')
-    price = models.FloatField(help_text='in CAD dollars $')
+    price = models.DecimalField(help_text='in CAD dollars $')
     created = models.DateTimeField(auto_now_add=True)
     updated = models.DateTimeField(auto_now=True)
